Candy-Bot_Backup/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.8.0
ffmpeg-python
python-dotenv
async_timeout
pytz
PyNaCl